ensure cleanup of prepdir even on error
We always create a prepdir to use its path in argparse, however, prior
to this, situations could occur, such as the program erroring out, that
would not clean up this prepdir.
Now this prepdir is always cleaned up, even when debugging mode is used,
which would leave the prepdir previously. If files need to be inspected
for debugging, an explicitly created prepdir has to be supplied going
forward.